Event cleanup, moved interactionCreate to own file
